MVR-200 Add auditing fields to Recipe entity
The Recipe class now implements IAuditable interface and has 4 new fields: createdDate, lastUpdatedDate, createdBy and lastUpdatedBy. This was added to keep track of who created the recipe and when it was last updated. Also, methods markAsCreated and markAsUpdated are added for easier set of these fields. Tests and other necessary classes on backend are adjusted according to those new changes.