MVR-294 FIXED bug where last_cooked was being overwritten on Recipe update
Remove unnecessary fields in RecipeServiceImpl

The code contains a commented block that shows a theoretical structure of the Recipe model that is not needed, thus cluttered the code. Unnecessary setter methods for "cooked" and "lastCooked" fields have been removed in the 'updateRecipe' function. This is done because the values for these fields are not updated through this method. Corresponding changes have been reflected in the RecipeServiceImplTest as well.