MVR-171 Integration Tests for Update Recipe
Refactor RestIntegrationTests and add new JSON test file

The RestIntegrationTests class has been refactored to be cleaner and more efficient. The previous version was functionally correct, but with this change, the code is now more readable which can greatly help with future debugging and modifications.

The main changes are extraction of PostRecipe into a separate method and refactoring of the testUpdateRecipe method, making it easier to understand and modify. The update tests start from 200 to maintain meaningful testing sequence.

A new JSON file chana_masala.json was also added to the test resources which can be utilized in future tests. This is to diversify our test cases and make sure our application can handle different data inputs.